C++ fixes for mac build

+++ src/common/Transforms.h
@@ -145,7 +145,7 @@ std::vector<T> localNormalize(
     int64_t curFrame = 0;
     // accumulate sum, sum^2 for computing mean, stddev
     for (auto i = b * perBatchSz; i < (b + 1) * perBatchSz; ++i) {
-      auto start = std::max(curFrame - rightCtxSize, 0L);
+      auto start = std::max(curFrame - rightCtxSize, int64_t(0));
       auto end = std::min(curFrame + leftCtxSize, frameSz - 1);
       for (int64_t j = start; j <= end; ++j) {
         sum[j] += in[i];
@@ -156,7 +156,7 @@ std::vector<T> localNormalize(
     // compute mean, stddev
     for (auto j = 0; j < frameSz; ++j) {
       int64_t N = (std::min(j + rightCtxSize, frameSz - 1) -
-                   std::max(j - leftCtxSize, 0L) + 1) *
+                   std::max(j - leftCtxSize, int64_t(0)) + 1) *
           perFrameSz;
       sum[j] /= N;
       sum2[j] /= N;
